+++ qiskit_qm_provider/__init__.py
@@ -48,12 +48,12 @@ from .primitives.qm_sampler import QMSamplerV2, QMSamplerOptions
 from .primitives.qm_estimator import QMEstimatorV2, QMEstimatorOptions
 from .backend.backend_utils import (
     get_measurement_outcomes,
-    add_basic_macros,
     assign_struct_with_table,
     dump_qua_script,
     get_qua_script,
     pack_register_to_int,
 )
+from .quam_macros import add_basic_macros
 from .backend.qua_circuit_compilation import (
     QuaCircuitCompilation,
     MeasurementOutcomeTable,